Web: Files: refactoring: created withLoadingCheck HOC
