+++ products/ASC.Files/Client/src/components/pages/Home/Section/Body/FilesRow/FilesRowContainer.js
@@ -1,42 +1,13 @@
-import React, { useEffect, useState } from "react";
+import React from "react";
 import { inject, observer } from "mobx-react";
 import RowContainer from "@appserver/components/row-container";
 import { Consumer } from "@appserver/components/utils/context";
 import SimpleFilesRow from "./SimpleFilesRow";
-import Loaders from "@appserver/common/components/Loaders";
-import { isMobile } from "react-device-detect";
 
-let loadTimeout = null;
+import withLoadingCheck from "../hoc/withLoadingCheck";
 
-const FilesRowContainer = ({ isLoaded, isLoading, filesList, tReady }) => {
-  const [inLoad, setInLoad] = useState(false);
-
-  const cleanTimer = () => {
-    loadTimeout && clearTimeout(loadTimeout);
-    loadTimeout = null;
-  };
-
-  useEffect(() => {
-    if (isLoading) {
-      cleanTimer();
-      loadTimeout = setTimeout(() => {
-        console.log("inLoad", true);
-        setInLoad(true);
-      }, 500);
-    } else {
-      cleanTimer();
-      console.log("inLoad", false);
-      setInLoad(false);
-    }
-
-    return () => {
-      cleanTimer();
-    };
-  }, [isLoading]);
-
-  return !isLoaded || (isMobile && inLoad) || !tReady ? (
-    <Loaders.Rows />
-  ) : (
+const FilesRowContainer = ({ filesList }) => {
+  return (
     <Consumer>
       {(context) => (
         <RowContainer
@@ -57,12 +28,10 @@ const FilesRowContainer = ({ isLoaded, isLoading, filesList, tReady }) => {
   );
 };
 
-export default inject(({ auth, filesStore }) => {
-  const { filesList, isLoading } = filesStore;
+export default inject(({ filesStore }) => {
+  const { filesList } = filesStore;
 
   return {
     filesList,
-    isLoading,
-    isLoaded: auth.isLoaded,
   };
-})(observer(FilesRowContainer));
+})(withLoadingCheck(observer(FilesRowContainer)));
