+++ products/ASC.Files/Client/src/components/pages/Home/Section/Body/hoc/withLoadingCheck.js
@@ -1,0 +1,51 @@
+import React, { useEffect, useState } from "react";
+import { observer, inject } from "mobx-react";
+import { isMobile } from "react-device-detect";
+
+import Loaders from "@appserver/common/components/Loaders";
+
+let loadTimeout = null;
+export default function withLoadingCheck(WrappedComponent, type) {
+  const WithLoadingCheck = (props) => {
+    const { tReady, firstLoad, isLoaded, isLoading } = props;
+    const [inLoad, setInLoad] = useState(false);
+
+    const cleanTimer = () => {
+      loadTimeout && clearTimeout(loadTimeout);
+      loadTimeout = null;
+    };
+
+    useEffect(() => {
+      if (isLoading) {
+        cleanTimer();
+        loadTimeout = setTimeout(() => {
+          console.log("inLoad", true);
+          setInLoad(true);
+        }, 500);
+      } else {
+        cleanTimer();
+        console.log("inLoad", false);
+        setInLoad(false);
+      }
+
+      return () => {
+        cleanTimer();
+      };
+    }, [isLoading]);
+
+    return firstLoad || !isLoaded || (isMobile && inLoad) || !tReady ? (
+      <Loaders.Rows />
+    ) : (
+      <WrappedComponent {...props} />
+    );
+  };
+
+  return inject(({ auth, filesStore }) => {
+    const { firstLoad, isLoading } = filesStore;
+    return {
+      firstLoad,
+      isLoaded: auth.isLoaded,
+      isLoading,
+    };
+  })(observer(WithLoadingCheck));
+}
